Release 2.0
In this release a menu structure has been built in.
From the main screen press * to reach it.
In every menu # takes you back to the pevious screen.

We can now also store settings in ATMEGA's EEPROM.
A boot up question has been added asking wether you would like to restore the settings.
(default is yes; be aware NOT to allow or choose that the first time because the EEPROM does not contain any settings yet.